Add container protocol for generic constraints
Similar to swift protocols such as `Collection`, `Sequence`, etc.

This provides a protocol with the main functionality of `IdentifiedArray` for use in generic constraints for types which can support some protocol IdentifiedArray conforms to. (see example in tests using `T: Collection`), allowing types to provide specialized functionality for given containers.

Before continuing on the todo list below, I wanted to provide this simple sample to see if this is something that would be considered for merge. I am currently using a similar protocol in a project by extending `IdentifiedArray` to conform outside of this package, but figured it may be useful for others.

TODO:
- [ ] Consensus on protocol name
- [ ] Add other public functions (eg: initializers)
- [ ] Move documentation?
- [ ] Verify if `@` attributes are needed/desired in protocol
- [ ] Protocol type documentation